Add startTimer method to allow async profiling.
When the profile method is used, ids are used to link the start and
end of the profiling.  Multiple simultaneous calls can result in the
profiling occuring in the wrong order.  E.g.:

function a(){
  winston.profile('X');
  setTimeout(function(){
    winston.profile('X');
  }, 1000);
}

a();
a();

Will result in both profiles reporting instantaneous results.

startTimer is a new logger function which returns an object
corresponding to a specific timing.  done is called to finish the
timer and log the duration E.g.:

timer = winston.startTimer()
setTimeout(function(){
  timer.done("Logging message");
}, 1000);