remove call to clone(meta) before calling custom formatter.
Without this fix, passing in any object with a circular reference
as a meta parameter will cause the program to crash with a:

    RangeError: Maximum call stack size exceeded

as common.clone does not handle circular references.  This cloning
is done before the users custom formatter function is called, so
even if they define a format function that returns 'foo', the app
will crash before their function is called, making it difficult to
track down the issue.

I tried modifying the line to call cycle.detect() as happens earlier
in the function, but couldn't get that to work, so I just removed the
call to clone.  Meta is already cloned above in the function so I don't
think it is needed.