Update file.js - objectMode writable stream
Appending this.eol explicitly to the result that common.log returns makes the output always become a string. That eliminates the possibility of piping js objects to the writable stream (which could be extremely handy for some cases). The proposed change won't break the code as the only case for the output not to be an object would be if this.stringify is explicitly rewritten to return objects (i.e. transport.stringify = function(obj){return obj})