Added ability to provide dynamic label.
I wanted to be able to provide a dynamic stacktrace for log functions. The workaround used to be just creating a new winston object wrapped in my own "debug/info/etc..." functions. This change allows you to pass in a function to the label option that will be executed at runtime and expected to yeild the dynamic label.