Added `isLevelEnabled(string)` & `isXXXEnabled()`
Make logger expose functions to check if a particular level is enabled.
This could be useful if either message or parameter creation is an
expensive operation, which should be avoided when log level is disabled.

Logger will contain `isLevelEnabled(level: string)` function and
multiple `isXXXEnabled()` where "XXX" is the capitalized name of the
configured log level. E.g. `isDebugEnabled()`, `isInfoEnabled()`, etc.