fix: assign log levels to un bound functions (#1483)
* fix: assign log levels to un bound functions 

binding custom log level functions to `this` with arrow functions causes unintentional behavior. 
the expectation is that they should be callable the same way as logger.log

if you extend a logger instance `logger.log` will be called on `logger` whereas `logger.warn` will be called on `logger.prototype` 

this fixes #1482 but I believe it should be treated as a bug. I doubt blocking extension of only part of the surface area of the instance was intended. 

this sets the base for efficient child loggers that only extend write to add attributes.

* test: test for custom method prototype binding

* style