Add gz extension on renaming first file
This problem verifies when both zippedArchive and tailable are true
In winston 3.1 you modified the feature brought by zippedArchive option: now a gzipped stream is created for the current file as well. In this case while rolling (tailable), you have to rename <file>.gz to <file>1.gz.
Relative documentation should be updated too.