Enables hrTime for profiler
In a NodeJS environment, [`process.hrtime()`][1] can be used to get the
current time with nanosecond resolution. We can use that to increase
the precision of the profiler.

I added a check for whether `hrTime` is available to the `Profiler` to
be able to use it. Since the logged message will now be a decimal with
fractions of milliseconds, I added an optional boolean argument to keep
backwards compatibility. This new argument must be set to `true`
explicitly to try to use `hrTime`. If `hrTime` is not available, it
will fall back to ms resolution using `Date`.

I added the same new optional boolean to the `logger`'s `startTimer()`
function. However, I don't see a way how to add it in a backwards
compatible way to the `logger`'s `profile` method. Is it ok to check if
`args` contains a boolean and in that case enable `useHrTime`?

However, I refactored `profile` a bit to use the same `Profiler` class
that `startTimer()` usees. The existing test ensure that it still works
as expected. I think I also found and fixed a bug regarding the
callback check. It should also be done if `this.profilers[id]` is not
set. And it should check the overall second argument (first in `args`),
not the second to last argument. This is now handled that way. Is it
correct now?

I also updated the README.
I also updated the type definition.

The following example program compares the profiler with and without
`hrTime` enabled (in a NodeJS environment):

```js
const winston = require('winston');
const { execSync } = require('child_process');

const logger = winston.createLogger({
  level: 'info',
  transports: [
    new winston.transports.Console({ format: winston.format.json() }),
  ],
});

let profiler = logger.startTimer();
execSync('ls -al');
profiler.done({hrTime: false});

profiler = logger.startTimer(true);
execSync('ls -al');
profiler.done({hrTime: true});
```

Output:
```js
{"hrTime":false,"level":"info","durationMs":11}
{"hrTime":true,"level":"info","durationMs":13.036486}
```

[1]: https://nodejs.org/docs/latest-v10.x/api/process.html#process_process_hrtime_time