fix: update exception handler tests to pass with jest
Previously the exception handler tests were written in a way that
a child node process was spawned by invoking a script that would
throw an unhandled exception. This exception would invoke the exception
handler and result in an invocation of `process.exit()`. Because
of this these lines weren't being included in the coverage report
resulting in what appeared to be a coverage regression.

I opted to refactor the implementation to hopefully clarify what is being tested
and how. The helpers while useful for the implementation, added a level
of obfuscation that did not make it immediately obvious what the test
did exactly.